revert pppoe discovery behaviour change
commit 1c082acf77e6256d9e47a3c9490b7615d591e67b changed the behaviour of
pppoe discovery such that if pppoe_verbose is != 0 then there's a hard-coded
delay of 5 seconds between receiving PADO and sending PADR.

This delay makes it impossible to complete a connection when the AC is sensibly
expecting a timely reply.
In normal circumstances the full PADI, PADO, PADR, PADS completes in sub-second
timeframes. The problem commit drastically changes the behaviour.

So remove the things that prevent exiting the loop when pppoe_verbose != 0
and a valid PADO has been received and prevent it continuing to loop on
errors.

Note that the problem commit seems to have been done with the intention of
some other behaviour for pppoe-discovery.c which I won't attempt to address
here as:
 1. breaking behaviour of the plugin is bad
 2. overloading the meaning of pppoe_verbose is just wrong as it guarantees
    that existing configs will get broken
 3. I don't believe there's any value to, or place for, pppoe-discovery